Fix signed/unsigned comparison warnings

+++ EEPROMex.h
@@ -87,7 +87,7 @@ class EEPROMClassEx
 	{
 		if (!isWriteOk(address+items*sizeof(T))) return 0;
 		unsigned int i;
-		for (i = 0; i < items; i++) 
+		for (i = 0; i < (unsigned int)items; i++)
 			readBlock<T>(address+(i*sizeof(T)),value[i]);
 		return i;
 	}
@@ -102,7 +102,7 @@ class EEPROMClassEx
 	{	
 		if (!isWriteOk(address+items*sizeof(T))) return 0;
 		unsigned int i;
-		for (i = 0; i < items; i++) 
+		for (i = 0; i < (unsigned int)items; i++)
 			  writeBlock<T>(address+(i*sizeof(T)),value[i]);
 		return i;
 	}
@@ -119,7 +119,7 @@ class EEPROMClassEx
 		int writeCount=0;
 		if (!isWriteOk(address+items*sizeof(T))) return 0;
 		unsigned int i;
-		for (i = 0; i < items; i++) 
+		for (i = 0; i < (unsigned int)items; i++)
 			  writeCount+= updateBlock<T>(address+(i*sizeof(T)),value[i]);
 		return writeCount;
 	}
@@ -129,7 +129,7 @@ class EEPROMClassEx
 		int writeCount=0;
 		if (!isWriteOk(address+sizeof(value))) return 0;
 		const byte* bytePointer = (const byte*)(const void*)&value;
-		for (unsigned int i = 0; i < sizeof(value); i++) {
+		for (unsigned int i = 0; i < (unsigned int)sizeof(value); i++) {
 			if (read(address)!=*bytePointer) {
 				write(address, *bytePointer);
 				writeCount++;
